refactor: move helper allocation functions into util.go
